refactor(database_utils): move function
