+++ apps/backend/src/fitness/resolver.rs
@@ -27,12 +27,11 @@ use serde::{Deserialize, Serialize};
 use services::FileStorageService;
 use slug::slugify;
 use traits::AuthProvider;
-use utils::GraphqlRepresentation;
+use utils::{user_measurements_list, GraphqlRepresentation};
 
 use crate::{
     app_utils::{
-        add_entity_to_collection, entity_in_collections, ilike_sql, user_by_id,
-        user_measurements_list, workout_details,
+        add_entity_to_collection, entity_in_collections, ilike_sql, user_by_id, workout_details,
     },
     background::{ApplicationJob, CoreApplicationJob},
 };
