+++ crates/utils/database/src/lib.rs
@@ -1,0 +1,23 @@
+use async_graphql::Result;
+use database_models::{prelude::UserMeasurement, user_measurement};
+use fitness_models::UserMeasurementsListInput;
+use sea_orm::{ColumnTrait, DatabaseConnection, EntityTrait, QueryFilter, QueryOrder, QueryTrait};
+
+pub async fn user_measurements_list(
+    db: &DatabaseConnection,
+    user_id: &String,
+    input: UserMeasurementsListInput,
+) -> Result<Vec<user_measurement::Model>> {
+    let resp = UserMeasurement::find()
+        .apply_if(input.start_time, |query, v| {
+            query.filter(user_measurement::Column::Timestamp.lte(v))
+        })
+        .apply_if(input.end_time, |query, v| {
+            query.filter(user_measurement::Column::Timestamp.gte(v))
+        })
+        .filter(user_measurement::Column::UserId.eq(user_id))
+        .order_by_asc(user_measurement::Column::Timestamp)
+        .all(db)
+        .await?;
+    Ok(resp)
+}
