+++ crates/utils/src/lib.rs
@@ -1,2 +1,3 @@
 pub use application_utils::*;
 pub use common_utils::*;
+pub use database_utils::*;
